Enable VP9 codec in standalone sender and receiver
This patch adds support for the VP9 video codec by modifying the former
streaming VP8 encoder to a general streaming VPX encoder that handles
both VP8 and VP9. It also adds a command line argument to the sender
to allow choosing which codec to use.

Change-Id: Ia254df071c37dddcbe8ea136aaaee47bc8db80a2
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/3028642
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Commit-Queue: mark a. foltz <mfoltz@chromium.org>