[Open Screen] Remove calls to python from DEPS.
Most of the scripts we use in DEPS already have correct shebang lines,
and don't need to be invoked with python3.

The exception is clang-update.py, which we download and run, and we
can't reliably set it to be executable on Windows.

This also fixes curlish.py to be py3 compatible.

Bug: b/195416694
Change-Id: If119386666364a56e833d0f6207437f1f5454d23
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/3253431
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Jordan Bayles <jophba@chromium.org>