[Cast Streaming] Support multiple senders
Currently, the ReceiverSession only supports having one Sender connected
at a time. However, this assumption doesn't match the current behavior
of receivers, which allow for multiple senders to connect and
renegotiate in the same session.

This patch removes the assumption that we are only connected to one
Sender in the ReceiverSession, and allows for multiple sender IDs in the
same session.

Bug: b/193416438
Change-Id: I84af8d6bfb9a9fbe47d9b07d10f17a101f17e8d1
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/3322101
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>