Avoid tautological comparison to 0 when char is unsigned.
The `char` type can be either signed or unsigned. If it is unsigned,
comparing to see if it's >= 0 is always true, and generates a warning
with -Wtautological-unsigned-zero-compare.

In this case, the code that invalidly (on unsigned-char builds)
compares to >= 0 is dead code, so we remove the function entirely.

Bug: chromium:1287175
Change-Id: Iff5c570bbdbfa7b294a1762b94e68f88f4cb8912
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/3399046
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Mark Foltz <mfoltz@chromium.org>