Add certificate verifier API
This change adds a certificate verifier API for use by cast sockets, so
that embedders can substitute their own certificate verification code.
This change provides the existing verifier functionality in the form of
an implementation of this API using boringssl, which can still be linked
by embedders that don't need an alternative.

Bug: b/216372726
Change-Id: I3b31a58c7cefd5cf7450dfdac4756b548d3596de
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/3415816
Reviewed-by: Mark Foltz <mfoltz@chromium.org>