[Open Screen] Remove third_party/chromium_quic.
This directory is a fork of the Chromium QUIC code from 2018 and has not
been kept in sync with upstream.  It requires non-trivial maintenance to
keep it compiling, is only used for OSP demo apps, and probably does not
interoperate with up-to-date QUIC implementations.

It will be replaced with a proper QUICHE port in the future.  For now,
to reduce code maintenance costs better to remove it.

Bug: b/169447969
Change-Id: I2970ee1a0ef43b3cc142e3b02351f0d2c13842b9
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/3573094
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Commit-Queue: Mark Foltz <mfoltz@chromium.org>