[Cast Streaming] Fix standalone sender
Currently, the standalone sender does not successfully connect
to a running standalone receiver due to it not properly setting
the developer certificate as the root for the cast trust store.

This change propagates the cast trust store to the sender socket
factory, allowing authentication to proceed.

Change-Id: I35d5c324026629c227a3cfad6faf646dd1fbec52
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/3685132
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>