[Open Screen] Update visibility rules for new cast_channel location.
We are moving the downstream cast_channel code so we need to add the new
location to GN visibilty rules in Open Screen.

Note: It's not great to hard code Chromium paths in Open Screen; is
there is a better way to control this?

Bug: 1358623
Change-Id: Ie00b8af8646b1e56e5da78d88ef1c86a374b7d9f
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/3919792
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Commit-Queue: Mark Foltz <mfoltz@chromium.org>