[Cast Streaming] fix timeout bug
Currently if we get a reply timeout it can cause a crash
if the callback method results in the session messenger getting
torn down or its replies get erased.

This patch resolves this issue by removing the callback from the
replies vector before executing its callback, meaning that the
callback execution is the last reference to the session messenger.

Bug: b/250957657
Change-Id: Ida174770e3f8133c166123de1b9d4ae9cd8680e3
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/3933581
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>