[Open Screen] Finish adding GN visibility and header checks.
This patch adds GN visibility and public checks to cast/protocol,
cast/sender, cast/receiver, and cast/streaming.

It also updates the default file-level visibility to be "./*", which
means targets in the same folder or sub-folders can depend on targets
in that file.  This simplifies the visibility lists in some cases.

This CL finishes the work to add these checks for the current
codebase.  The last item will be to add documentation on how to
maintain these checks.

NOTE #1: It would be nice to allowlist certain targets from visibility
checks like the standalone sender and receiver and gn_all.  However
not aware of a way to do that.

NOTE #2: Some Cast Streaming headers are used only for unit tests in
Chromium or not used at all.  Filed crbug.com/1378386 to investigate
making these headers private.

Bug: 1358623,1378386,b/154090565
Change-Id: I00918375ff41ea665d07f6f617a0332f36bb21de
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/3980173
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Reviewed-by: Ryan Keane <rwkeane@google.com>
Commit-Queue: Mark Foltz <mfoltz@chromium.org>