[Infra] Update cast E2E bot
Previously, we were blocked on getting a cast standalone end to end
test bot running because the bots do not have the required packages.

This patch is a first step towards fixing this, since when we are cross
compiling we have all of the packages necessary as part of the sysroot,
meaning that we can at least compile the standalone code.

In a future patch we will have to update and see if we can get the end
to end tests running.

Bug: b/269515650
Change-Id: I647aaab4f805bc3e6fa50e4d35885f72536c67db
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/4257318
Reviewed-by: Mark Foltz <mfoltz@chromium.org>