[Open Screen] Convert mDNS TXT records to hold octet data.
TXT records typically hold character data, however, it is simpler
to treat it as binary data (octets) in the internal implementation and
provide conversion functions in a higher level API (DNS-SD).

Conversion to passing the TXT data as ByteView also removes several
unnecessary copies and conversions.

This also allows the removal of the one usage of absl::Span in the
discovery module.

Change-Id: Ie5c073c3a798b88b73056a81e472c3b335856562
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/4546027
Commit-Queue: Mark Foltz <mfoltz@chromium.org>
Reviewed-by: Jordan Bayles <jophba@chromium.org>