[OpenScreen] Replace TaskRunner* with TaskRunner& in discovery/ directory
TaskRunner can be passed by reference instead of pointer as it should
never be null. The change will be made gradually. This CL makes this
change in discovery/ directory. There will be other CLs for other
directories and the directories that have been changed are as follows:
[1] osp/: https://chromium-review.googlesource.com/c/openscreen/+/4599931
[2] cast/: https://chromium-review.googlesource.com/c/openscreen/+/4602778

Bug: b/281739775
Change-Id: I6498b461b2652ee5daf5efe55edbf4d751b7224e
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/4606448
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Mark Foltz <mfoltz@chromium.org>