[OpenScreen] Replace TaskRunner* with TaskRunner& in util/ directory
TaskRunner can be passed by reference instead of pointer as it should
never be null. The change will be made gradually. This CL makes this
change in util/ directory. There will be the last CL for platform/
directory and the directories that have been changed are as follows:
[1] osp/: https://chromium-review.googlesource.com/c/openscreen/+/4599931
[2] cast/: https://chromium-review.googlesource.com/c/openscreen/+/4602778
[3] discovery/: https://chromium-review.googlesource.com/c/openscreen/+/4606448


Bug: b/281739775
Change-Id: I8f21b60f5978976be18c5930907f043acc4cdffb
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/4608512
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Mark Foltz <mfoltz@chromium.org>