[Open Screen] Resolve ubsan failure in statistics_analyzer_unittest.cc
Event structs were being initialized with a reciever_timestamp of
Clock::time_point::min(), which results in underflow in the unit test.

Refactor event construction in the unit test to set a valid mock
timestamp and reduce the amount of boilerplate code in the test.

Fixed: 1498251
Change-Id: I8986fd679f3195f4c4d8cee6ae153248ab50197f
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/5031771
Reviewed-by: Jordan Bayles <jophba@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>