[Build] Move openscreen_include_dirs
This config is part of the build and needs to still exist after
we migrate to using Chrome's build system. This patch breaks out
the config migration so we can consume the updated location in
Chrome in advance of landing the main build patch.

Change-Id: Iacf7a977ea23bb7e850636ba19b20a6e443abafb
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/5287834
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Jordan Bayles <jophba@chromium.org>