Remove cast_allow_developer_certificate build flag
This flag was used by Chrome, but as of crrev.com/c/5407786, it is not
used. It should be able to be removed.

This CL has the effect of always enabling the flag for the standalone
receiver and standalone sender binaries.

Bug: b:320187114
Test: Chrome and openscreen targets build

Change-Id: Ie04de6e55ff9498ca4c28cbdf33dcc7ebaeba07a
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/5445045
Commit-Queue: Mark Foltz <mfoltz@chromium.org>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Nathan Hebert <nhebert@chromium.org>