[Open Screen] Fix uninitialized memory in Linux interface query.
Some of the code to query for network interfaces on Linux was failing
msan because of structs that are not fully initialized.  This CL fixes
the cases caught by msan.

The rest of the platform impl should also be audited, tracked under
the linked bug.

Bug: 349649079
Change-Id: I89772b3e405a018c8ab6d92e137a96e4c408c68b
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/5660248
Commit-Queue: Mark Foltz <mfoltz@chromium.org>
Reviewed-by: Takumi Fujimoto <takumif@chromium.org>