[OpenScreen] Don't create ProtocolConnection in OnCryptoHandshakeComplete
Currently, a ProtocolConnection is created in OnCryptoHandshakeComplete
and passed to embedder through OnConnectionOpened. However, in some
cases, the embedder may just want to establish the connection and
create the ProtocolConnection later. Besides, this makes the
ProtocolConnectionClient::Connect interface confusing.

The main changes in this CL are:
  1. Rename ConnectionRequestCallback to ConnectRequestCallback to keep
     align with ConnectRequest, and rename interfaces for it.
  2. Don't create ProtocolConnection in OnCryptoHandshakeComplete.
     Instead, embedder should create ProtocolConnection itself after
     OnConnectSucceed is called.

Bug: 347197917
Change-Id: Ibf194f5f78a81d8f8ec9b68bc2e57bd3f48f4772
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/5733162
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Wei4 Wang <wei4.wang@intel.com>