[OpenScreen] Include authentication token in DNS TXT records
According to the spec[1], the authentication token(at) of the
advertising agent should be included in DNS TXT records.

Once this CL is merged, we can generate authentication token on the
server side and send it to client side. Then client can use it in
authentication.

[1] https://w3c.github.io/openscreenprotocol/#discovery

Bug: none
Change-Id: Ied5dbcb9f12d5bd3e920346422e3832f00b0d4e7
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/5768956
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Commit-Queue: Wei4 Wang <wei4.wang@intel.com>