[OpenScreen] Work around CDDL optional member bug for PresentationConnectionCloseEvent
According to [1], optional member in CDDL is only supoorted when it has
a value or it doesn't have a value but it is the last member. This
CL works around this bug for PresentationConnectionCloseEvent by moving
the optional member to the last one.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/openscreen/src/tools/cddl/codegen.cc;l=1523

Bug: 322772635
Change-Id: I5655858ddb4d4539db97aef2d28634a01e147e7e
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/5820004
Commit-Queue: Wei4 Wang <wei4.wang@intel.com>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>