[osp] Remove friendly_name from ServiceInfo.
The friendly_name is not advertised as part of the mDNS service info
per the OSP spec.  Instead the agent display name is returned as part
of the agent metadata.

https://www.w3.org/TR/openscreen-application/#metadata

This removes the fn key from the mDNS TXT constants and the
friendly_name field from the ServiceInfo struct in the OSP
implementation to match the spec.

Change-Id: I57b74bc24f7da344c377cf3a1a9d03d7ebea5de0
Reviewed-on: https://chromium-review.googlesource.com/c/openscreen/+/6086360
Commit-Queue: Mark Foltz <mfoltz@chromium.org>
Reviewed-by: Jordan Bayles <jophba@chromium.org>